# fix: optimize related jobs query performance with hash-based randomization